// packages/js/data/src/payment-settings/action-types.ts
export enum ACTION_TYPES {
	GET_PAYMENT_PROVIDERS_REQUEST = 'GET_PAYMENT_PROVIDERS_REQUEST',
	GET_PAYMENT_PROVIDERS_SUCCESS = 'GET_PAYMENT_PROVIDERS_SUCCESS',
	GET_PAYMENT_PROVIDERS_ERROR = 'GET_PAYMENT_PROVIDERS_ERROR',
	UPDATE_PROVIDER_ORDERING = 'UPDATE_PROVIDER_ORDERING',
	SET_IS_ELIGIBLE = 'SET_IS_ELIGIBLE',
}